change in Dish model to set float, ReviewWebService calculates dish avg

+++ src/main/java/model/Dish.java
@@ -116,7 +116,7 @@ public class Dish {
 		return ratingAvg;
 	}
 
-	public void setRatingAvg(int ratingAvg) {
+	public void setRatingAvg(float ratingAvg) {
 		this.ratingAvg = ratingAvg;
 	}
